+++ elasticmagic-query-filters/src/commonMain/kotlin/dev/evo/elasticmagic/qf/SimpleQueryFilter.kt
@@ -3,6 +3,8 @@ package dev.evo.elasticmagic.qf
 import dev.evo.elasticmagic.query.Bool
 import dev.evo.elasticmagic.query.QueryExpression
 
+class SimpleQueryValue(val name: String, val expr: QueryExpression, val title: String = "")
+
 open class SimpleQueryFilter(
     name: String? = null,
     val values: List<SimpleQueryValue> = emptyList(),
