--- elasticmagic-query-filters/src/commonMain/kotlin/dev/evo/elasticmagic/qf/SimpleQueryValue.kt
@@ -1,6 +1,0 @@
-package dev.evo.elasticmagic.qf
-
-import dev.evo.elasticmagic.query.QueryExpression
-
-
-class SimpleQueryValue(val name: String, val expr: QueryExpression, val title: String = "")
